+++ CHANGELOG.md
@@ -1,3 +1,8 @@
 # Changelog of JEI LOTR
+
+## 1.1.0-BETA
+* New faction crafting tables will now be discovered automatically
+* Fixed that some recipes were not shown for faction crafting tables
+
 ## 1.0.0-BETA:
 * First release for LOTR Renewed-1.1 and jei-1.15.2-6.0.0.4